Add offset

+++ src/wz_file.rs
@@ -80,7 +80,7 @@ impl WzFile {
     }
 
     pub fn parse_root_directory(&mut self) -> Result<ArcWzNode, Box<dyn std::error::Error>> {
-        let offset = get_version_offset(*self.reader.file_start.borrow(), self.version);
+        let offset = get_version_offset(*self.reader.file_start.borrow() as usize, self.version);
 
         let node = parse_directory(self.name.clone(), &self.reader.clone(), offset)?;
